requirements.txt.txt
torch
torchvision
numpy
Pillow
tqdm
scikit-learn
opencv-python
matplotlib
